Rename fdd to fddplan
