+++ dedisp/cpp/include/fddplan.hpp
